+++ src/ClearFeatureActions/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.1.0] - 2025-04-25
+### Added
+- INotification
+- INotificationHandler<TNotification>
+- INotificationPublisher<TNotification>
+- NotificationPublisher<TNotification>
+### Changed
+- ServiceCollectionExtensions
+
 ## [1.0.1] - 2025-04-22
 ### Changed
 - Updated `AddFeatureActions` service extension method